[util] Move view options into view-options.hh
